Use Connection for Playlists
The assumption that only a few playlists (<1k) was made when originally
using an array. This puts the responsible of sorting on the client side.
For most other queries, data is sorted on the server.

This will be useful for finding the only the most recently added
playlists and the most recently played playlists.