Added timeAdded, Removed Artist.songs, Added SortedConnectionQuery
Added timeAdded. This information will need to be stored for the
RECENTLY_ADDED sort.

Removed Artist.songs. This is hard to implement on the server. It brings
little value without shuffle.

Added SortedConnectionQuery. There are two types of places connections
are used. One is with search and sorting and the other is without either
(just simple pagination). Separated these into two different argument
types.