Removed RECENTLY_PLAYED and MOST_PLAYED
This requires defining what playing means for albums, artists and songs.

We were considering counting a play on an album, artist and playlist
when a song was played from the list, but this would show plays on
playlists which weren't plays of a playlist, just songs in the playlist.

To get around this the playSong mutation could specify where a song is
being played from (playlist, album, artist). This would give the server
enough information to find the most played albums, playlist, artists and
songs but it would introduce data storage complexity.

This will be added back in the future with an enhanced
Mutation.playSong.