Moved Test Runner to Package
The tests are exposed through the forte-test-api command. This allows
for jest to have full control over the testing context at the loss of
modularity. This is ok. The modularity was used to switch the client to
a local, mock, resolver. The mock resolver should be exposed over HTTP
like other resolvers.