Added Lerna
There are pros to mono-repos and cons to mono-repos. I've decided that
the benifits of colocated code, versioned together, tested together and
evolving together outweighs the cost of handling publishing and adding
lerna.

This change also removed publishing to npm from CircleCI. The process is
easy (lerna publish) and will be non-frequent after the constructing
phase.