Removed Play Count from Album, Artist, Playlist
It only really makes sense for the last played time to be stored and
exposed for these groups because the song count can be greater than the
album play count when the song is played from a playlist many times then
viewed in the album.

Also removed MOST_PLAYED sort. This is because there is no longer a
notion of most played album, artist and playlist. There can still be a
most played song. Maybe a smart playlist can be used to display songs
from most played to least played. This is not a priority.

Closes #34