Added Information About Remote Resources
Previously, the URL could have been any url. It should be a relative url
because the server may not know how the client is accesing it. The
server could be behind a proxy. The only guarantee is that the client
can access resources relative to a configured server path which it can
access.